Move logic from SolidityTracer to ErrorInferrer
